+++ tests/testthat/test-coerce.R
@@ -1,23 +1,27 @@
 test_that("Autodetect", {
   samples <- rep(c("a", "b", "c", "d", "e"), each = 3)
   groups <- rep(c("A", "B", "C"), each = 5)
+  dates <- sample(1400:1451, 15, TRUE)
   tpq <- sample(1301:1400, 15, TRUE)
   taq <- sample(1451:1500, 15, TRUE)
 
   X <- matrix(data = sample(0:10, 75, TRUE), nrow = 15, ncol = 5)
-  Y <- data.frame(samples = samples, groups = groups, tpq = tpq, taq = taq, X)
+  Y <- data.frame(samples = samples, groups = groups, dates = dates,
+                  tpq = tpq, taq = taq, X)
 
   options(arkhe.autodetect = FALSE)
   Z <- as_count(Y)
   expect_equal(get_samples(Z), rownames(Y))
   expect_equal(get_groups(Z), character(0))
-  expect_equal(get_dates(Z), data.frame(tpq = integer(0), taq = integer(0)))
+  expect_equal(get_dates(Z), integer(0))
+  expect_equal(get_terminus(Z), data.frame(tpq = integer(0), taq = integer(0)))
 
   options(arkhe.autodetect = TRUE)
   Z <- as_count(Y)
   expect_equal(get_samples(Z), samples)
   expect_equal(get_groups(Z), groups)
-  expect_equal(get_dates(Z), data.frame(tpq = tpq, taq = taq),
+  expect_equal(get_dates(Z), dates)
+  expect_equal(get_terminus(Z), data.frame(tpq = tpq, taq = taq),
                ignore_attr = TRUE)
 
   Y$label <- LETTERS[1:15]
@@ -29,7 +33,7 @@ test_that("Autodetect", {
   expect_equal(dim(A), dim(X) + c(0, 1))
 
   Y <- data.frame(samples = character(0), groups = character(0),
-                  tpq = integer(0), taq = integer(0))
+                  dates = integer(0), tpq = integer(0), taq = integer(0))
   expect_error(as_count(Y))
 })
 # matrix =======================================================================
@@ -111,11 +115,12 @@ test_that("DataMatrix > long", {
   counts <- as(cts, "CountMatrix")
 
   A <- as_long(counts, factor = TRUE, reverse = FALSE)
-  expect_equal(dim(A), c(50, 7))
+  expect_equal(dim(A), c(50, 8))
   expect_s3_class(A$row, "factor")
   expect_s3_class(A$column, "factor")
+  expect_type(A$dates, "integer")
   expect_type(A$tpq, "integer")
-  expect_type(A$tpq, "integer")
+  expect_type(A$taq, "integer")
 
   B <- as_long(counts, factor = TRUE, reverse = TRUE)
   expect_equal(rev(levels(A$row)), levels(B$row))
@@ -125,33 +130,5 @@ test_that("DataMatrix > features", {
   freq <- as_composition(cts)
   feat <- as_features(freq)
 
-  expect_equal(dim(feat), c(5, 14))
-})
-# Autodetect ===================================================================
-test_that("Autodetect", {
-  spl <- LETTERS[1:5]
-  grp <- c("a", "a", "b", "b", "c")
-
-  cts <- matrix(sample(0:100, 50, TRUE), ncol = 10)
-  cts <- as.data.frame(cts)
-  cts$sample <- spl
-  cts$group <- grp
-
-  options(arkhe.autodetect = TRUE)
-  counts <- as_count(cts)
-  expect_equal(get_samples(counts), spl)
-  expect_equal(get_groups(counts), grp)
-
-  freq <- as_composition(cts)
-  expect_equal(get_samples(freq), spl)
-  expect_equal(get_groups(freq), grp)
-
-  incid <- as_incidence(cts)
-  expect_equal(get_samples(incid), spl)
-  expect_equal(get_groups(incid), grp)
-
-  options(arkhe.autodetect = FALSE)
-  counts <- as_count(cts)
-  expect_equal(get_samples(counts), rownames(cts))
-  expect_equal(get_groups(counts), character(0))
+  expect_equal(dim(feat), c(5, 15))
 })