+++ tests/testthat/test-subset.R
@@ -64,10 +64,13 @@ test_that("Subset extra slots", {
   set_groups(cts) <- rep(c("A", "B"), each = 10)
   set_samples(cts) <- rep(c("X", "Y"), times = 10)
   set_dates(cts) <- 1:20
+  set_tpq(cts) <- 1:20
+  set_taq(cts) <- 21:40
 
   tmp <- cts[1:10, ]
-  dts <- data.frame(tpq = 1:10, taq = 1:10, row.names = paste0("row", 1:10))
   expect_identical(get_groups(tmp), rep("A", 10))
   expect_identical(get_samples(tmp), rep(c("X", "Y"), times = 5))
-  expect_identical(get_dates(tmp), dts)
+  expect_identical(get_dates(tmp), 1:10)
+  expect_identical(get_tpq(tmp), 1:10)
+  expect_identical(get_taq(tmp), 21:30)
 })